src/backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
mssql-django==1.5
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
